+++ examples/gke_shared_vpc/main.tf
@@ -14,18 +14,12 @@
  * limitations under the License.
  */
 
-locals {
-  credentials_file_path = var.credentials_path
-}
-
 provider "google" {
-  credentials = file(local.credentials_file_path)
-  version     = "~> 3.30"
+  version = "~> 3.30"
 }
 
 provider "google-beta" {
-  credentials = file(local.credentials_file_path)
-  version     = "~> 3.30"
+  version = "~> 3.30"
 }
 
 provider "null" {
@@ -37,13 +31,12 @@ provider "random" {
 }
 
 module "project-factory" {
-  source             = "../../"
-  random_project_id  = true
-  name               = "sample-gke-shared-project"
-  org_id             = var.org_id
-  billing_account    = var.billing_account
-  shared_vpc         = var.shared_vpc
-  activate_apis      = ["compute.googleapis.com", "container.googleapis.com", "cloudbilling.googleapis.com"]
-  credentials_path   = local.credentials_file_path
-  shared_vpc_subnets = var.shared_vpc_subnets
+  source               = "../../"
+  random_project_id    = true
+  name                 = "sample-gke-shared-project"
+  org_id               = var.org_id
+  billing_account      = var.billing_account
+  svpc_host_project_id = var.shared_vpc
+  activate_apis        = ["compute.googleapis.com", "container.googleapis.com", "cloudbilling.googleapis.com"]
+  shared_vpc_subnets   = var.shared_vpc_subnets
 }