+++ examples/project_services/main.tf
@@ -21,14 +21,25 @@ provider "google" {
   version = "~> 3.30"
 }
 
+provider "google-beta" {
+  version = "~> 3.38"
+}
+
 module "project-services" {
   source                      = "../../modules/project_services"
   project_id                  = var.project_id
   enable_apis                 = var.enable
-  disable_services_on_destroy = "true"
+  disable_services_on_destroy = true
 
   activate_apis = [
     "sqladmin.googleapis.com",
     "bigquery-json.googleapis.com",
   ]
+  activate_api_identities = [{
+    api = "healthcare.googleapis.com"
+    roles = [
+      "roles/healthcare.serviceAgent",
+      "roles/bigquery.jobUser",
+    ]
+  }]
 }