+++ test/fixtures/minimal/main.tf
@@ -44,8 +44,8 @@ module "project-factory" {
     "container.googleapis.com",
   ]
 
-  default_service_account     = "disable"
-  disable_services_on_destroy = "false"
+  default_service_account     = "DISABLE"
+  disable_services_on_destroy = false
 }
 
 // Add a binding to the container service robot account to test that the